fix: write tests for adapter-nostr-tools
