+++ packages/adapter-nostr-tools/package.json
@@ -29,6 +29,8 @@
   "scripts": {
     "prepack": "yarn build",
     "tsc": "tsc",
+    "test": "vitest run",
+    "testw": "vitest",
     "lint": "run-p tsc lint:*",
     "lint:format": "prettier --check --loglevel warn src/**/*.ts",
     "lint:js": "eslint --max-warnings 0 src/",
@@ -43,7 +45,8 @@
     "@nostr-fetch/kernel": "^0.9.0"
   },
   "devDependencies": {
-    "nostr-tools": "^1.3.0"
+    "nostr-tools": "^1.3.0",
+    "@nostr-fetch/testutil": "^0.9.0"
   },
   "peerDependencies": {
     "nostr-fetch": "^0.9.0",
